+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/PolicyAPIGroupClient.java
@@ -15,15 +15,9 @@
  */
 package io.fabric8.kubernetes.client;
 
-import io.fabric8.kubernetes.api.model.policy.v1beta1.PodDisruptionBudget;
-import io.fabric8.kubernetes.api.model.policy.v1beta1.PodDisruptionBudgetList;
-import io.fabric8.kubernetes.api.model.policy.v1beta1.PodSecurityPolicy;
-import io.fabric8.kubernetes.api.model.policy.v1beta1.PodSecurityPolicyList;
-import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.PolicyAPIGroupDSL;
-import io.fabric8.kubernetes.client.dsl.Resource;
-import io.fabric8.kubernetes.client.dsl.internal.policy.v1beta1.PodDisruptionBudgetOperationsImpl;
-import io.fabric8.kubernetes.client.dsl.internal.policy.v1beta1.PodSecurityPolicyOperationsImpl;
+import io.fabric8.kubernetes.client.dsl.V1PolicyAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.V1beta1PolicyAPIGroupDSL;
 import okhttp3.OkHttpClient;
 
 public class PolicyAPIGroupClient extends BaseClient implements PolicyAPIGroupDSL {
@@ -33,12 +27,12 @@ public class PolicyAPIGroupClient extends BaseClient implements PolicyAPIGroupDS
   }
 
   @Override
-  public MixedOperation<PodSecurityPolicy, PodSecurityPolicyList, Resource<PodSecurityPolicy>> podSecurityPolicies() {
-    return new PodSecurityPolicyOperationsImpl(httpClient, getConfiguration());
+  public V1PolicyAPIGroupDSL v1() {
+    return adapt(V1PolicyAPIGroupClient.class);
   }
 
   @Override
-  public MixedOperation<PodDisruptionBudget, PodDisruptionBudgetList, Resource<PodDisruptionBudget>> podDisruptionBudget() {
-    return new PodDisruptionBudgetOperationsImpl(httpClient, getConfiguration());
+  public V1beta1PolicyAPIGroupDSL v1beta1() {
+    return adapt(V1beta1PolicyAPIGroupClient.class);
   }
 }