+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/V1PolicyAPIGroupClient.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright (C) 2015 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.fabric8.kubernetes.client;
+
+import io.fabric8.kubernetes.api.model.policy.v1.PodDisruptionBudget;
+import io.fabric8.kubernetes.api.model.policy.v1.PodDisruptionBudgetList;
+import io.fabric8.kubernetes.client.dsl.MixedOperation;
+import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.kubernetes.client.dsl.V1PolicyAPIGroupDSL;
+import io.fabric8.kubernetes.client.dsl.internal.policy.v1.PodDisruptionBudgetOperationsImpl;
+import okhttp3.OkHttpClient;
+
+public class V1PolicyAPIGroupClient extends BaseClient implements V1PolicyAPIGroupDSL {
+  public V1PolicyAPIGroupClient() {
+    super();
+  }
+
+  public V1PolicyAPIGroupClient(OkHttpClient httpClient, final Config config) {
+    super(httpClient, config);
+  }
+
+  @Override
+  public MixedOperation<PodDisruptionBudget, PodDisruptionBudgetList, Resource<PodDisruptionBudget>> podDisruptionBudget() {
+    return new PodDisruptionBudgetOperationsImpl(httpClient, getConfiguration());
+  }
+}