+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/V1BatchAPIGroupDSL.java
@@ -15,11 +15,22 @@
  */
 package io.fabric8.kubernetes.client.dsl;
 
-public interface Evictable<B> {
+import io.fabric8.kubernetes.api.model.batch.v1.CronJob;
+import io.fabric8.kubernetes.api.model.batch.v1.CronJobList;
+import io.fabric8.kubernetes.api.model.batch.v1.Job;
+import io.fabric8.kubernetes.api.model.batch.v1.JobList;
+import io.fabric8.kubernetes.client.Client;
+
+public interface V1BatchAPIGroupDSL extends Client {
   /**
-   * Evicts resource, respecting {@link io.fabric8.kubernetes.api.model.policy.v1beta1.PodDisruptionBudget}
-   * @return value indicating object was evicted or not
-   * @throws io.fabric8.kubernetes.client.KubernetesClientException if an error occurs.
+   * DSL entrypoint for batch/v1 Job
+   * @return {@link MixedOperation} for Job resource
    */
-  B evict();
+  MixedOperation<Job, JobList, ScalableResource<Job>> jobs();
+
+  /**
+   * DSL entrypoint for batch/v1 CronJob
+   * @return {@link MixedOperation} for CronJob resource
+   */
+  MixedOperation<CronJob, CronJobList, Resource<CronJob>> cronjobs();
 }