+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/V1DiscoveryAPIGroupDSL.java
@@ -15,11 +15,10 @@
  */
 package io.fabric8.kubernetes.client.dsl;
 
-public interface Evictable<B> {
-  /**
-   * Evicts resource, respecting {@link io.fabric8.kubernetes.api.model.policy.v1beta1.PodDisruptionBudget}
-   * @return value indicating object was evicted or not
-   * @throws io.fabric8.kubernetes.client.KubernetesClientException if an error occurs.
-   */
-  B evict();
+import io.fabric8.kubernetes.api.model.discovery.v1.EndpointSlice;
+import io.fabric8.kubernetes.api.model.discovery.v1.EndpointSliceList;
+import io.fabric8.kubernetes.client.Client;
+
+public interface V1DiscoveryAPIGroupDSL extends Client {
+  MixedOperation<EndpointSlice, EndpointSliceList, Resource<EndpointSlice>> endpointSlices();
 }