+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/V1PolicyAPIGroupDSL.java
@@ -15,11 +15,14 @@
  */
 package io.fabric8.kubernetes.client.dsl;
 
-public interface Evictable<B> {
+import io.fabric8.kubernetes.api.model.policy.v1.PodDisruptionBudget;
+import io.fabric8.kubernetes.api.model.policy.v1.PodDisruptionBudgetList;
+import io.fabric8.kubernetes.client.Client;
+
+public interface V1PolicyAPIGroupDSL extends Client {
   /**
-   * Evicts resource, respecting {@link io.fabric8.kubernetes.api.model.policy.v1beta1.PodDisruptionBudget}
-   * @return value indicating object was evicted or not
-   * @throws io.fabric8.kubernetes.client.KubernetesClientException if an error occurs.
+   * DSL entrypoint for policy/v1 PodDisruptionBudget
+   * @return {@link MixedOperation} for PodDisruptionBudget
    */
-  B evict();
+  MixedOperation<PodDisruptionBudget, PodDisruptionBudgetList, Resource<PodDisruptionBudget>> podDisruptionBudget();
 }